+++ package.json
@@ -4,9 +4,10 @@
   "description": "Run Architect projects locally",
   "main": "src/index.js",
   "scripts": {
-    "test": "npm run lint && npm run test:unit",
-    "test:unit": "tape test/*-test.js | tap-spec",
-    "coverage": "istanbul cover tape test/*-test.js",
+    "test": "npm run lint && npm run test:unit && npm run test:integration",
+    "test:unit": "tape test/unit/*-test.js test/unit/**/*-test.js | tap-spec",
+    "test:integration": "tape test/integration/*-test.js | tap-spec",
+    "coverage": "istanbul cover tape test/unit/*-test.js test/unit/**/*-test.js",
     "lint": "eslint src --fix"
   },
   "author": "Brian LeRoux <b@brian.io>",
