+++ test/integration/mock/no-index/src/http/get-foo/.arc-config
