+++ test/integration/mock/normal/.arc
