+++ test/integration/mock/normal/src/events/ping/.arc-config
