+++ test/integration/mock/normal/src/events/ping/index.js
