+++ test/integration/mock/normal/src/http/get-index/.arc-config
