+++ test/integration/mock/normal/src/http/get-index/index.js
