+++ test/integration/mock/normal/src/queues/pong/.arc-config
