+++ test/integration/mock/normal/src/queues/pong/index.js
